Second Submission Module 2 Challenge
My initial submission was returned by the tester stating that they were receiving a complier error when running the code.  I did pull the .VBS code from the initial submission, upload it into the initial raw data Excel workbook and was able to run it without receiving any errors.  In Monday night I forwarded repository link to the instructional staff, and they too were able to run the code with no errors.  The instructional staff were not able to identify any potential compiling issues and were not able to replicate the error so they asked if the tester might be using an older version of Excel which may be causing the error.  To provide options for the tester to successfully run my code, I have submitted the code in .vbs, .bas and .txt versions.  Also, if the compiler error persists, I was encouraged to request screen shots of the error, as we can not replicate the error to identify potential solutions.  Thank you for your time and efforts!

Michael Leonard